application/src/: rename `endpoints/` to be `api/`
